create protected route

+++ backend/routes/Addmeal.js
@@ -77,7 +77,30 @@ router.post("/deletecustomrecipe/:id",async(req,res)=>{
         const { id } = req.params; 
         const existUser=await User.findById(userId);
         if(!existUser) return res.status(404).json({message:"User not found"});
+        const meal = await customrecipe.findOne({ _id: id, userRef:userId });
+        if (!meal) {
+            return res.status(404).json({ message: "Meal not found or unauthorized to delete" });
+        }
         const result = await customrecipe.findByIdAndDelete(id);
+        if (!result) {
+         return res.status(404).json({ message: 'Recipe not found' });
+        }
+        return res.status(200).json({message:"Recipe deleted successfully"});
+    } catch (error) {
+        res.status(500).json({message:error.message});
+    }
+})
+router.post("/deletemeal/:id",async(req,res)=>{
+    try {
+        const{userId}=req.body;
+        const { id } = req.params; 
+        const existUser=await User.findById(userId);
+        if(!existUser) return res.status(404).json({message:"User not found"});
+        const meal = await addmeal.findOne({ _id: id, userRef:userId });
+        if (!meal) {
+            return res.status(404).json({ message: "Meal not found or unauthorized to delete" });
+        }
+        const result = await addmeal.findByIdAndDelete(id);
         if (!result) {
          return res.status(404).json({ message: 'Recipe not found' });
         }
